admin and trainer search done

+++ js/AdminMealPlan.js
@@ -103,7 +103,7 @@ function appendMealSection(mealPlan) {
         // Find the #mealPlanName element within the card and get its text content
         let mealId = card.find("#mealId").text();
         console.log(mealId);
-        setDeleteModalContent( mealId);
+        setDeleteModalContent(mealId);
 
     })
 
@@ -123,7 +123,7 @@ function appendMealSection(mealPlan) {
 
 
 // meal plan save method
-$("#saveMeal").click(function (){
+$("#saveMeal").click(function () {
 
     let meal_id = $("#meal_id").val();
     let meal_name = $("#meal_name").val();
@@ -188,8 +188,7 @@ $("#saveMeal").click(function (){
 })
 
 
-
-$("#updateMeal").click(function (){
+$("#updateMeal").click(function () {
 
     let meal_id = $("#Update_meal_id").val();
     let meal_name = $("#Update_meal_name").val();
@@ -277,8 +276,6 @@ $("#updateMeal").click(function (){
 // })
 
 
-
-
 // method to set data to update modal text fields
 function setUpdateModalContent(mealPlanName, mealPlanDetails, calorie, mealId) {
     $("#Update_meal_id").val(mealId);
@@ -289,22 +286,28 @@ function setUpdateModalContent(mealPlanName, mealPlanDetails, calorie, mealId) {
 }
 
 // method to set data to delete modal text fields
-function setDeleteModalContent( mealId) {
+function setDeleteModalContent(mealId) {
 
-    $.ajax({
-        url: 'http://localhost:8080/api/v1/mealPlan/delete/' + mealId,
-        method: "DELETE",
-        success: function (response) {
-            console.log(response)
-            getAll();
-            alert("Meal Plan Deleted Successfully !!")
-            $('#deleteMealModal').data('bs.modal').hide();
-        },
+    var result = window.confirm("Do you want to proceed?");
+    if (result) {
+        $.ajax({
+            url: 'http://localhost:8080/api/v1/mealPlan/delete/' + mealId,
+            method: "DELETE",
+            success: function (response) {
+                console.log(response)
+                getAll();
+                alert("Meal Plan Deleted Successfully !!")
+                $('#deleteMealModal').data('bs.modal').hide();
+            },
 
-        error: function (jqXHR) {
-            console.log(jqXHR);
-        }
-    })
+            error: function (jqXHR) {
+                console.log(jqXHR);
+            }
+        })
+    } else {
+        alert("Your Meal Plan Is Safe !!")
+    }
+
 
 }
 
@@ -390,12 +393,11 @@ $("#memberComboBox").click(function () {
 
     })
 
-    })
-
+})
 
 
 // update user with mealPlan
-$("#assignMealPlanBtn").click(function (){
+$("#assignMealPlanBtn").click(function () {
     console.log(memId);
 
     console.log(memberEmail);
@@ -430,6 +432,91 @@ $("#assignMealPlanBtn").click(function (){
         }
     })
 
+})
+
+
+// meal plan search by name
+
+$("#SearchMeal").keyup(function () {
+
+    let text = $('#SearchMeal').val();
+    $.ajax({
+        url: 'http://localhost:8080/api/v1/mealPlan/searchMealByName',
+        method: 'GET',
+        dataType: 'json',
+        data: {partialName: text},   // Convert data to JSON string
+        success: function (response) {
+            console.log(response);
+            if ($("#SearchMeal").val() === "") {
+                $("#cardContainer").css("justifyContent","start")
+                getAll();
+
+            } else {
+                let cardContainer = $("#cardContainer");
+                cardContainer.empty();
+                $.each(response.data, function (index, mealPlan) {
+                    console.log(mealPlan);
+
+                    let card = `
+  <section class="mx-3 my-5" style="max-width: 20rem;">
+
+    <div id="card" class="card" >
+   
+      <div class="card-header px-4" style="background-color: #2d324a; color: white">
+      <p id="mealPlanName" class="mb-0" style="font-size: 1rem; font-weight: 400 !important;"><a>${mealPlan.planName}</a></p>
+     <p class="small mb-0">meal plan id:&nbsp;&nbsp;<span id="mealId">${mealPlan.mid}</span></p>
+</div>
+
+     <div class="dropdown position-absolute threeDots">
+                                     <a class="btn btn-secondary dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                                        <i class="fa fa-ellipsis-v" aria-hidden="true"></i>
+                                     </a>
+                                     <ul class="dropdown-menu">
+                                        <li><a id="edit"  class="dropdown-item edit" href="#" data-toggle="modal" data-target="#updateMealModal" >Edit</a></li>
+                                        <li><a class="dropdown-item delete" href="#" >Delete</a></li>
+                                        <li><a class="dropdown-item assign" href="#" data-toggle="modal" data-target="#assignModal">Assign</a></li>
+                                     </ul>
+                                </div>
+
+      <div class="card-body">
+
+      
+
+        <p id="mealPlanDetail" class="card-text">${mealPlan.planDetails}</p>
+
+        <hr class="my-4" />
+        <p  class="lead"><strong>Total calorie count : <span id="mealPlanCalorie">${mealPlan.calorieCount}</span> </strong></p>
+
+      </div>
+
+    </div>
+  </section>
+  
+  
+`
+                    $("#cardContainer").append(card);
+
+
+                });
+
+            }
+        },
+        error: function (jqXHR, textStatus, errorThrown) {
+            console.error(jqXHR.responseText);  // Log the response text for debugging
+           if (jqXHR.data == null) {
+               $("#cardContainer").empty();
+
+               let card = `
+ <img src="https://cdn.dribbble.com/users/1242216/screenshots/9326781/media/6384fef8088782664310666d3b7d4bf2.png" alt="no" width="620px">
+
+  
+`
+               let cardContainer=$("#cardContainer");
+               cardContainer.css("justify-content","center")
+               cardContainer.append(card);
+           }
+        }
+    });
 })
